# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "AgreementMetrics": {
        "properties": {
          "agree": {
            "default": 0,
            "title": "Agree",
            "type": "integer"
          },
          "agreement": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agreement"
          },
          "kappa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kappa"
          },
          "n": {
            "default": 0,
            "title": "N",
            "type": "integer"
          }
        },
        "title": "AgreementMetrics",
        "type": "object"
      },
...
      "Criterion": {
        "properties": {
          "baseline": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline"
          },
          "criterion": {
            "title": "Criterion",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "met": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Met"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          },
          "now": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Now"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "id",
          "criterion",
          "metric",
          "target"
        ],
        "title": "Criterion",
        "type": "object"
      },
...
      "DimensionSummary": {
        "properties": {
          "confusion": {
            "$ref": "#/components/schemas/Confusion"
          },
          "f1": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "F1"
          },
          "n": {
            "title": "N",
            "type": "integer"
          },
          "precision": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precision"
          },
          "recall": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recall"
          }
        },
        "required": [
          "precision",
          "recall",
          "f1",
          "n",
          "confusion"
        ],
        "title": "DimensionSummary",
        "type": "object"
      },
...
      "EvalItem": {
        "description": "One labelled item with its evaluator verdict (planted-failure gallery and drill-down).",
        "properties": {
          "aspect_ratio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aspect Ratio"
          },
          "brief_id": {
            "title": "Brief Id",
            "type": "string"
          },
          "dimensions": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Dimensions",
            "type": "object"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "title": "Evidence",
            "type": "array"
          },
          "geography_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "From data/golden/briefs.yaml",
            "title": "Geography Code"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "image_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Id"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "label": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "boolean"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "label_source": {
            "title": "Label Source",
            "type": "string"
          },
          "mutation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mutation"
          },
          "origin": {
            "enum": [
              "natural",
              "planted"
            ],
            "title": "Origin",
            "type": "string"
          },
          "outcomes": {
            "additionalProperties": {
              "anyOf": [
                {
                  "enum": [
                    "tp",
                    "fp",
                    "fn",
                    "tn"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Outcomes",
            "type": "object"
          },
          "product_id": {
            "description": "Golden product key (P1..P5)",
            "title": "Product Id",
            "type": "string"
          },
          "product_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Name"
          },
          "required_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Text"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The golden run that produced this image (natural items)",
            "title": "Run Id"
          },
          "season": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season"
          },
          "set": {
            "enum": [
              "nat",
              "final",
              "plant"
            ],
            "title": "Set",
            "type": "string"
          },
          "source_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Image Url"
          },
          "split": {
            "enum": [
              "calibration",
              "heldout"
            ],
            "title": "Split",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "verdict": {
            "enum": [
              "pass",
              "fail",
              "unverified"
            ],
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "id",
          "origin",
          "set",
          "brief_id",
          "product_id",
          "split",
          "image_id",
          "image_url",
          "source_image_url",
          "mutation",
          "label_source",
          "label",
          "verdict",
          "dimensions",
          "outcomes",
          "evidence"
        ],
        "title": "EvalItem",
        "type": "object"
      },
...
      "EvalSummary": {
        "description": "The latest eval report (architecture API contract `EvalSummary`, plus the PRD criteria).",
        "properties": {
          "after_repair_pass_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Repair Pass Rate"
          },
          "agreement_by_dimension": {
            "additionalProperties": {
              "additionalProperties": {
                "$ref": "#/components/schemas/AgreementMetrics"
              },
              "type": "object"
            },
            "description": "Human agreement per image set (nat|final) and dimension, with Cohen's kappa",
            "title": "Agreement By Dimension",
            "type": "object"
          },
          "comparison": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GoldenComparison"
              },
              {
                "type": "null"
              }
            ],
            "description": "v1 -> v2: human and evaluator pass rates per dimension (incl. composition) and evaluator P/R per version; on the newest version's report"
          },
          "cost_per_approved_ad": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Per Approved Ad"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/Criterion"
            },
            "title": "Criteria",
            "type": "array"
          },
          "dataset_version": {
            "title": "Dataset Version",
            "type": "string"
          },
          "dry_run": {
            "title": "Dry Run",
            "type": "boolean"
          },
          "evaluator_version": {
            "title": "Evaluator Version",
            "type": "string"
          },
          "first_attempt_pass_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Attempt Pass Rate"
          },
          "git_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Sha"
          },
          "golden_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Golden dataset version (v1 | v2) of this report",
            "title": "Golden Version"
          },
          "hemisphere": {
            "$ref": "#/components/schemas/RateMetric"
          },
          "human_agreement": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Human Agreement"
          },
          "human_agreement_n": {
            "title": "Human Agreement N",
            "type": "integer"
          },
          "human_kappa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Human Kappa"
          },
          "judge_stability_flip_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Judge Stability Flip Rate"
          },
          "known_good": {
            "$ref": "#/components/schemas/RateMetric"
          },
          "native_text_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Native Text Rate"
          },
          "p50_latency_to_approved_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P50 Latency To Approved Ms"
          },
          "p95_latency_to_approved_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P95 Latency To Approved Ms"
          },
          "per_dimension": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DimensionSummary"
            },
            "title": "Per Dimension",
            "type": "object"
          },
          "per_split": {
            "additionalProperties": {
              "additionalProperties": {
                "$ref": "#/components/schemas/DimensionSummary"
              },
              "type": "object"
            },
            "title": "Per Split",
            "type": "object"
          },
          "pipeline": {
            "$ref": "#/components/schemas/PipelineMetrics",
            "description": "Pipeline quality on the golden runs: mean_repairs, status_counts (runs by status), pass rates, cost and latency"
          },
          "planted": {
            "items": {
              "$ref": "#/components/schemas/PlantedClassMetrics"
            },
            "title": "Planted",
            "type": "array"
          },
          "provenance": {
            "$ref": "#/components/schemas/EvalProvenance"
          },
          "report_file": {
            "title": "Report File",
            "type": "string"
          },
          "report_id": {
            "format": "uuid",
            "title": "Report Id",
            "type": "string"
          },
          "stability": {
            "$ref": "#/components/schemas/StabilityMetrics"
          },
          "targets": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Targets",
            "type": "object"
          },
          "text_exact_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Exact Rate"
          },
          "warnings": {
            "description": "FAKE DRY RUN, missing labels, stale snapshot/report and the report's own",
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "report_id",
          "created_at",
          "evaluator_version",
          "dataset_version",
          "git_sha",
          "dry_run",
          "report_file",
          "per_dimension",
          "per_split",
          "human_agreement",
          "human_agreement_n",
          "human_kappa",
          "first_attempt_pass_rate",
          "after_repair_pass_rate",
          "native_text_rate",
          "text_exact_rate",
          "cost_per_approved_ad",
          "p50_latency_to_approved_ms",
          "p95_latency_to_approved_ms",
          "judge_stability_flip_rate",
          "planted",
          "known_good",
          "criteria",
          "targets",
          "counts"
        ],
        "title": "EvalSummary",
        "type": "object"
      },
...
      "GoldenComparison": {
        "description": "v1 (baseline, orchestrator-3) vs v2 (ADR-007, orchestrator-4), in version order.",
        "properties": {
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionSummary"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "versions"
        ],
        "title": "GoldenComparison",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_EvalItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EvalItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque; pass as `cursor` to get the next page.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[EvalItem]",
        "type": "object"
      },
...
      "RateMetric": {
        "properties": {
          "n": {
            "default": 0,
            "title": "N",
            "type": "integer"
          },
          "ok": {
            "default": 0,
            "title": "Ok",
            "type": "integer"
          },
          "rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate"
          }
        },
        "title": "RateMetric",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ad-studio",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/evals/items": {
      "get": {
        "description": "Labelled items of the latest report with their verdicts (planted-failure gallery and the\nTP/FP/FN/TN drill-down; `outcome` needs the positive class FAIL).",
        "operationId": "evals_eval_items",
        "parameters": [
          {
            "in": "query",
            "name": "origin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "natural",
                    "planted"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          },
          {
            "in": "query",
            "name": "dimension",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "technical",
                    "text",
                    "product",
                    "context",
                    "composition"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dimension"
            }
          },
          {
            "in": "query",
            "name": "outcome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "tp",
                    "fp",
                    "fn",
                    "tn"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "`next_cursor` of the previous page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`next_cursor` of the previous page",
              "title": "Cursor"
            }
          },
          {
            "description": "Golden dataset version (v1 | v2); default: the newest report of any version",
            "in": "query",
            "name": "golden_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^v[0-9]{1,3}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Golden dataset version (v1 | v2); default: the newest report of any version",
              "title": "Golden Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_EvalItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Eval Items",
        "tags": [
          "evals"
        ]
      }
    },
    "/v1/evals/summary": {
      "get": {
        "description": "The latest evaluator report: per-dimension precision/recall/F1, human agreement, pipeline\npass rates, text exactness, $/ad, latency and every PRD criterion with its target.",
        "operationId": "evals_eval_summary",
        "parameters": [
          {
            "in": "query",
            "name": "evaluator_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Evaluator Version"
            }
          },
          {
            "description": "Golden dataset version (v1 | v2); default: the newest report of any version",
            "in": "query",
            "name": "golden_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^v[0-9]{1,3}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Golden dataset version (v1 | v2); default: the newest report of any version",
              "title": "Golden Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Eval Summary",
        "tags": [
          "evals"
        ]
      }
    },
...
  }
}